requirements.txt
mcp
PyYAML
